Refactor and move the context classes
* Rename the configuration context class to bring it inline with the other classes
* Move the context classes into a new package
* Clean up the code slightly and improve log output
* Remove dependencies on deprecated classes
