+++ grails-app/conf/application.groovy
@@ -1,6 +1,8 @@
+import grails.plugins.hibernate.search.context.HibernateSearchConfigurationInterceptor
+
 dataSources {
 
 	dataSource {
-		configClass = grails.plugins.hibernate.search.HibernateSearchConfigurationInterceptor
+		configClass = HibernateSearchConfigurationInterceptor
 	}
 }
